refactor: moving textToActor to index
